backend/requirements.txt
Django==5.2
django-cors-headers==4.7.0
pillow==11.2.1
djangorestframework==3.16.0
django-debug-toolbar==5.2.0
reportlab==4.4.1
qrcode==8.2
psycopg2==2.9.10
djoser==2.3.1
djangorestframework-simplejwt==5.5.0